+++ CHANGELOG.md
@@ -25,3 +25,11 @@
 -   `custom/plugins/StrixNLUxUpgrades/src/Resources/views/storefront/page/checkout/summary.html.twig`
 -   `custom/plugins/StrixNLUxUpgrades/src/Resources/views/storefront/page/checkout/summary/summary-pretotal.html.twig`
 -   `custom/plugins/StrixNLUxUpgrades/src/Resources/views/storefront/page/checkout/summary/summary-discount.html.twig`
+### 2025-09-07 14:54 — 20e2e2bb
+- Add feature to show discount in cart summary
+
+**Changed files:**
+
+- `custom/plugins/StrixNLUxUpgrades/src/Core/Checkout/Cart/Processor/CartDiscountProcessor.php`
+- `custom/plugins/StrixNLUxUpgrades/src/Resources/config/services.xml`
+